+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 All notable changes to this project will be documented in this file.
 
 
+
+## [0.2.2] - 2019-01-25
+
+### Fixed
+- In **Eval** string properties can now contain the '='
+- Message and Header encoding, support for larger get and put
+
+### Changed
+- Encoding of properties
+
 ## [0.2.1] - 2019-01-17
 
 ### Changed